// src/app/core/api/models/vigencias-update.ts
/* tslint:disable */
export interface VigenciasUpdate {
  estado: string;
  fechaActualiza?: string;
  fechaCreacion?: string;
  fechaFinal: string;
  fechaInicial: string;
  idParConcepto?: number;
  idVigencia: number;
  observaciones: string;
  periodo: string;
  seRepite: string;
  tipoPeriodo: string;
  usuarioActualiza?: string;
  usuarioCreacion?: string;
}